requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
sqlalchemy==2.0.35
pydantic==2.8.2
python-multipart==0.0.9
aiofiles==24.1.0
python-dotenv==1.0.1
pytest==8.3.2
httpx==0.27.0
